Az-204_AzureFunctions_Netflix/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
python-dotenv==1.0.0